chore: convert public key to bytes
