+++ contracts/shade/src/events.rs
@@ -1,4 +1,4 @@
-use soroban_sdk::{contractevent, Address, Env};
+use soroban_sdk::{contractevent, Address, BytesN, Env};
 
 #[contractevent]
 pub struct InitalizedEvent {
@@ -94,14 +94,14 @@ pub fn publish_merchant_verified_event(env: &Env, merchant_id: u64, status: bool
 #[contractevent]
 pub struct MerchantKeySetEvent {
     pub merchant: Address,
-    pub key: soroban_sdk::String,
+    pub key: BytesN<32>,
     pub timestamp: u64,
 }
 
 pub fn publish_merchant_key_set_event(
     env: &Env,
     merchant: Address,
-    key: soroban_sdk::String,
+    key: BytesN<32>,
     timestamp: u64,
 ) {
     MerchantKeySetEvent {
