+++ contracts/shade/src/shade.rs
@@ -7,7 +7,7 @@ use crate::errors::ContractError;
 use crate::events;
 use crate::interface::ShadeTrait;
 use crate::types::{ContractInfo, DataKey, Invoice, InvoiceFilter, Merchant, MerchantFilter, Role};
-use soroban_sdk::{contract, contractimpl, panic_with_error, Address, Env, String, Vec};
+use soroban_sdk::{contract, contractimpl, panic_with_error, Address, BytesN, Env, String, Vec};
 
 #[contract]
 pub struct Shade;
@@ -86,11 +86,11 @@ impl ShadeTrait for Shade {
         invoice_component::get_invoice(&env, invoice_id)
     }
 
-    fn set_merchant_key(env: Env, merchant: Address, key: String) {
+    fn set_merchant_key(env: Env, merchant: Address, key: BytesN<32>) {
         merchant_component::set_merchant_key(&env, &merchant, &key);
     }
 
-    fn get_merchant_key(env: Env, merchant: Address) -> String {
+    fn get_merchant_key(env: Env, merchant: Address) -> BytesN<32> {
         merchant_component::get_merchant_key(&env, &merchant)
     }
