+++ contracts/shade/src/tests/mod.rs
@@ -1,4 +1,5 @@
 pub mod test;
 pub mod test_accepted_tokens;
 pub mod test_invoice;
+pub mod test_merchant_key;
 pub mod test_upgrade;
