+++ contracts/shade/src/tests/test_invoice.rs
@@ -48,70 +48,6 @@ fn assert_latest_invoice_event(
 }
 
 #[test]
-fn test_set_merchant_key_success() {
-    let (env, client, contract_id, _admin) = setup_test();
-
-    let merchant = Address::generate(&env);
-    client.register_merchant(&merchant);
-
-    let key = String::from_str(&env, "GC3L4...EXAMPLE");
-    client.set_merchant_key(&merchant, &key);
-    let events = env.events().all();
-
-    assert_eq!(client.get_merchant_key(&merchant), key);
-
-    assert!(
-        events.len() > 0,
-        "No events captured immediately after key set!"
-    );
-
-    let (event_contract_id, _topics, data) = events.get(events.len() - 1).unwrap();
-    assert_eq!(event_contract_id, contract_id.clone());
-
-    let data_map: Map<Symbol, Val> = data.try_into_val(&env).expect("Data should be a Map");
-
-    let merchant_val = data_map
-        .get(Symbol::new(&env, "merchant"))
-        .expect("Should have merchant field");
-    let key_val = data_map
-        .get(Symbol::new(&env, "key"))
-        .expect("Should have key field");
-
-    let merchant_in_event: Address = merchant_val.try_into_val(&env).unwrap();
-    let key_in_event: String = key_val.try_into_val(&env).unwrap();
-
-    assert_eq!(merchant_in_event, merchant.clone());
-    assert_eq!(key_in_event, key.clone());
-}
-
-#[test]
-fn test_update_merchant_key() {
-    let (env, client, _contract_id, _admin) = setup_test();
-
-    let merchant = Address::generate(&env);
-    client.register_merchant(&merchant);
-
-    let key1 = String::from_str(&env, "KEY_ONE");
-    client.set_merchant_key(&merchant, &key1);
-    assert_eq!(client.get_merchant_key(&merchant), key1);
-
-    let key2 = String::from_str(&env, "KEY_TWO");
-    client.set_merchant_key(&merchant, &key2);
-    assert_eq!(client.get_merchant_key(&merchant), key2);
-}
-
-#[test]
-#[should_panic(expected = "HostError: Error(Contract, #11)")]
-fn test_get_non_existent_key() {
-    let (env, client, _contract_id, _admin) = setup_test();
-
-    let merchant = Address::generate(&env);
-    client.register_merchant(&merchant);
-
-    client.get_merchant_key(&merchant);
-}
-
-#[test]
 fn test_create_and_get_invoice_success() {
     let (env, client, contract_id, _admin) = setup_test();
